Move the HidApi implemenatation into each backend
Our use of HidApiBackend works a bit like a trait by failing at compile time if
a bckend has not implemented it. But it's not as clear-cut as having a specific
list of functions in a block.
